refactor(query): split query pipeline modules
